Estilo

+++ src/app/administrador-list/administrador-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Administrador } from '../administrador';
 import { AdministradorService } from '../administrador.service'; 
+import { DialogoService } from '../dialogo.service';
 
 @Component({
   selector: 'app-administrador-list',
@@ -13,7 +14,8 @@ export class AdministradorListComponent implements OnInit {
   administradores!: Administrador[];
 
   constructor(private administradorServicio: AdministradorService,
-    private router: Router) { }
+    private router: Router,
+    private dialogService: DialogoService) { }
 
   ngOnInit(): void {
     this.getAdministradores();
@@ -32,10 +34,21 @@ export class AdministradorListComponent implements OnInit {
     this.router.navigate(['actualizarAdministrador',id]);
   }
 
-  deleteAdministrador(id: number){
-    this.administradorServicio.deleteAdministrador(id).subscribe(data => {
-      console.log(data);
-      this.getAdministradores();
-    });
+  deleteAdministrador(id: number){ 
+    /*this.dialogService.abrirDialogoConfirmacion('¿Desea Borrar este Administrador?')
+    .afterClosed().subscribe(res =>{
+      if(res){
+        this.administradorServicio.deleteAdministrador(id).subscribe(data => {
+          console.log(data);
+          this.getAdministradores();
+        });
+      }
+    });*/
+    if(confirm('Desea eliminar el Administrador?')){
+      this.administradorServicio.deleteAdministrador(id).subscribe(data => {
+        console.log(data);
+        this.getAdministradores();
+      });
+    }
   }
 }
